requirements.txt
pyaudio
vosk
opencv-python
pillow
requests
numpy
duckduckgo-search
sentence-transformers
diffusers
transformers
accelerate
torch
pygame
webrtcvad
pyautogui
psutil
python-dotenv
keyboard